Move `CLanguage` and `LoadLanguageIndexfile` to localization
The `CLanguage` class and `LoadLanguageIndexfile` function are more appropriately located in the file for the other localization functions.
