+++ src/game/localization.h
@@ -4,8 +4,25 @@
 #define GAME_LOCALIZATION_H
 
 #include <base/system.h> // GNUC_ATTRIBUTE
+
 #include <engine/shared/memheap.h>
+
+#include <string>
 #include <vector>
+
+class CLanguage
+{
+public:
+	CLanguage() = default;
+	CLanguage(const char *pName, const char *pFileName, int Code) :
+		m_Name(pName), m_FileName(pFileName), m_CountryCode(Code) {}
+
+	std::string m_Name;
+	std::string m_FileName;
+	int m_CountryCode;
+
+	bool operator<(const CLanguage &Other) const { return m_Name < Other.m_Name; }
+};
 
 class CLocalizationDatabase
 {
@@ -27,6 +44,7 @@ class CLocalizationDatabase
 		bool operator==(const CString &Other) const { return m_Hash == Other.m_Hash && m_ContextHash == Other.m_ContextHash; }
 	};
 
+	std::vector<CLanguage> m_vLanguages;
 	std::vector<CString> m_vStrings;
 	CHeap m_StringsHeap;
 	int m_VersionCounter;
@@ -34,6 +52,9 @@ class CLocalizationDatabase
 
 public:
 	CLocalizationDatabase();
+
+	void LoadIndexfile(class IStorage *pStorage, class IConsole *pConsole);
+	const std::vector<CLanguage> &Languages() const { return m_vLanguages; }
 
 	bool Load(const char *pFilename, class IStorage *pStorage, class IConsole *pConsole);
